+++ src/Akeneo/Connectivity/Connection/front/src/shared/user/index.ts
@@ -1,5 +1,6 @@
-import {User as UserInterface} from './user.interface';
+import {useUser} from './use-user';
 import {UserContext} from './user-context';
+import {User as UserInterface} from './user.interface';
 import {UserButtons} from './UserButtons';
 
-export {UserContext, UserInterface, UserButtons};
+export {UserContext, UserInterface, UserButtons, useUser};