+++ src/Akeneo/Connectivity/Connection/front/tests/src/webhook/model/EventSubscriptionLogFilters.test.tsx
@@ -11,6 +11,7 @@ test('it compares successfully the default filters values', () => {
                 EventSubscriptionLogLevel.ERROR,
             ],
             text: '',
+            dateTime: {},
         })
     ).toBeTruthy();
 });
@@ -20,6 +21,7 @@ test('it fails to compare modified filters values with the default ones', () => 
         isSameAsDefaultFiltersValues({
             levels: [EventSubscriptionLogLevel.INFO],
             text: '',
+            dateTime: {},
         })
     ).toBeFalsy();
 
@@ -32,6 +34,7 @@ test('it fails to compare modified filters values with the default ones', () => 
                 EventSubscriptionLogLevel.ERROR,
             ],
             text: 'search value',
+            dateTime: {},
         })
     ).toBeFalsy();
 });