+++ src/Akeneo/Pim/Enrichment/Bundle/Product/ComputeAndPersistProductCompletenesses.php
@@ -14,7 +14,7 @@ use Akeneo\Pim\Enrichment\Component\Product\Query\SaveProductCompletenesses;
  */
 class ComputeAndPersistProductCompletenesses
 {
-    private const CHUNK_SIZE = 10000;
+    private const CHUNK_SIZE = 1000;
 
     /** @var CompletenessCalculator */
     private $completenessCalculator;