+++ src/Akeneo/Pim/Structure/Bundle/Resources/workspaces/settings-ui/src/apps/index.ts
@@ -1,2 +1,3 @@
+export * from './AttributeGuidelinesApp';
 export * from './AttributeGroupsApp';
 export * from './LocalesApp';