+++ src/Akeneo/Pim/Structure/Bundle/Resources/workspaces/settings-ui/src/components/index.ts
@@ -1,4 +1,5 @@
 export * from './buttons';
 export * from './datagrids';
+export * from './inputs';
 export * from './providers';
 export * from './shared';